Only remove zshrc if oh-my-zsh backed it up

+++ install.sh
@@ -53,6 +53,8 @@ echo "Installing zsh"
 . "$DOTFILES_DIR/install/zsh_install.sh"
 . "$DOTFILES_DIR/install/oh-my-zsh-install.sh"
 
-rm ~/.zshrc
-mv ~/.zshrc.pre-oh-my-zsh ~/.zshrc
+if ! test -f .zshrc.pre-oh-my-zsh; then
+  rm ~/.zshrc
+  mv ~/.zshrc.pre-oh-my-zsh ~/.zshrc
+fi
